project_mini/face_recognition_project/requirements.txt
opencv-python==4.6.0.66
Pillow==10.0.0
face_recognition==1.3.0
  